Chef and Stock Prices
Chef wants to buy a stock whose price was S rupees when the market opened. He will buy the stock if and only if its price is in the range [A,B]. The price of the stock has changed by C% by the time he was trying to buy the stock. Will he be able to buy the stock?

Sample 1:

Input

3
100 93 108 7
100 94 100 -7
183 152 172 -17

Output

Yes
No
No